Supporting nodmax and mayhem
gen8randombattlenodmax is said to be supported in the first array but it isn't. Mayhem another popular format uses the same movepools as gen8randombattlemayhem, so support for that is requested as well.  

Scarlett/Violet is coming in less than a week though so these changes may be considered irrelevant though....